- refactor withModuleConfig to advanced sample - removed withConsoleForwarding from samples
