+++ src/mono/sample/wasm/browser-advanced/index.html
