--- src/mono/sample/wasm/browser-native/main.js
@@ -1,28 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-import { dotnet, exit } from './dotnet.js'
-
-function displayMeaning(meaning) {
-    document.getElementById("out").innerHTML = `${meaning}`;
-}
-
-try {
-    const { setModuleImports } = await dotnet
-        .withConsoleForwarding()
-        .withElementOnExit()
-        .create();
-
-    setModuleImports("main.js", {
-        Sample: {
-            Test: {
-                displayMeaning
-            }
-        }
-    });
-
-    await dotnet.run();
-}
-catch (err) {
-    exit(2, err);
-}
