+++ src/mono/sample/wasm/browser-threads/main.js
@@ -1,19 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-import { dotnet } from './dotnet.js'
-
-function wasm_exit(exit_code, reason) {
-    /* Set result in a tests_done element, to be read by xharness in runonly CI test */
-    const tests_done_elem = document.createElement("label");
-    tests_done_elem.id = "tests_done";
-    tests_done_elem.innerHTML = exit_code.toString();
-    if (exit_code) tests_done_elem.style.background = "red";
-    document.body.appendChild(tests_done_elem);
-
-    if (reason) console.error(reason);
-    console.log(`WASM EXIT ${exit_code}`);
-}
+import { dotnet, exit } from './dotnet.js'
 
 let progressElement = null;
 
@@ -65,33 +53,9 @@ function onInputValueChanged(exports, inputElement) {
 try {
     const inputElement = document.getElementById("inputN");
     const { setModuleImports, getAssemblyExports, runMain } = await dotnet
-        .withModuleConfig({
-            // This whole 'withModuleConfig' is for demo purposes only.
-            // It is prefered to use specific 'with***' methods instead. 
-            // Only when such method is doesn't exist, fallback to moduleConfig.
-
-            onConfigLoaded: (config) => {
-                // This is called during emscripten `dotnet.wasm` instantiation, after we fetched config.
-                console.log('user code Module.onConfigLoaded');
-                // config is loaded and could be tweaked before the rest of the runtime startup sequence
-                config.environmentVariables["MONO_LOG_LEVEL"] = "debug"
-            },
-            preInit: () => { console.log('user code Module.preInit'); },
-            preRun: () => { console.log('user code Module.preRun'); },
-            onRuntimeInitialized: () => {
-                console.log('user code Module.onRuntimeInitialized');
-                // here we could use API passed into this callback
-                // Module.FS.chdir("/");
-            },
-            onDotnetReady: () => {
-                // This is called after all assets are loaded.
-                console.log('user code Module.onDotnetReady');
-            },
-            postRun: () => { console.log('user code Module.postRun'); },
-        })
+        .withEnvironmentVariable("MONO_LOG_LEVEL", "debug")
         .create();
 
-    console.log('user code after dotnet.create');
     setModuleImports("main.js", {
         Sample: {
             Test: {
@@ -107,7 +71,7 @@ try {
     inputElement.addEventListener("change", onInputValueChanged(exports, inputElement));
 
     let exit_code = await runMain(assemblyName, []);
-    wasm_exit(exit_code);
+    exit(exit_code);
 } catch (err) {
-    wasm_exit(2, err);
+    exit(2, err);
 }
